+++ tools/http-test-helper/src/main/java/org/enso/shttp/test_helpers/CrashingTestHandler.java
@@ -1,6 +1,8 @@
-package org.enso.shttp;
+package org.enso.shttp.test_helpers;
 
 import com.sun.net.httpserver.HttpExchange;
+import org.enso.shttp.SimpleHttpHandler;
+
 import java.io.IOException;
 
 public class CrashingTestHandler extends SimpleHttpHandler {
