+++ tools/http-test-helper/src/main/java/org/enso/shttp/test_helpers/GenerateDataLinkHandler.java
@@ -1,6 +1,7 @@
-package org.enso.shttp;
+package org.enso.shttp.test_helpers;
 
 import com.sun.net.httpserver.HttpExchange;
+import org.enso.shttp.SimpleHttpHandler;
 
 import java.io.IOException;
